Move elevon servos to the output rail on the NANOWII in FLYING_WING mode.
At least on Felix Nessen's NanoWii and Hobby King's NanoWii - both of which
suggest that board definition NANOWII should be used - Arduino pins 10 and 5
map to the second and third channels on the output rail that provides power
and ground pins next to each signal pin.

FLYING_WING mode uses SERVO_1 and SERVO_2 as elevons, which by default
are mapped to Arduino pins A0 and A1, which are accessible on these boards,
but lack the convenient power and ground pins. This commit re-maps the pins
for SERVO_1 and SERVO_2 (and SERVO_5 and SERVO_7) to the mentioned pins
on the output rail in this situation.

This change is probably only beneficial to you if you use one of the
mentioned boards AND want to fly in elevon mode AND you choose to use
pin headers and standard servo connectors instead of soldering
directly to the board.

AIRPLANE mode seems to map aileron, elevator and rudder to the second,
third and fourth channels (Arduino pins 10, 5, and 6) on the output rail
as desired, regardless of this change.

The throttle channel seems to map to the first channel on the output
rail in both AIRPLANE mode and FLYING_WING mode.

http://flyduino.net/documents/NanoWii_manual.pdf
http://www.hobbyking.com/hobbyking/store/__22322__MultiWii_NanoWii_ATmega32U4_Micro_Flight_Controller_USB_GYRO_ACC.html
http://multiwii.googlecode.com/svn/trunk/Doc/Diagrams/7%20Connection%20diagram%20flying%20wing.png